Avoid Z axis homing when laser has no focus

+++ MK/module/motion/planner.h
@@ -105,7 +105,7 @@ typedef struct {
     float laser_ppm; // pulses per millimeter, for pulsed and raster firing modes
     unsigned long laser_duration; // laser firing duration in microseconds, for pulsed and raster firing modes
     long steps_l; // step count between firings of the laser, for pulsed firing mode
-    int laser_intensity; // Laser firing instensity in clock cycles for the PWM timer
+    float laser_intensity; // Laser firing instensity in clock cycles for the PWM timer
     #if ENABLED(LASER_RASTER)
       unsigned char laser_raster_data[LASER_MAX_RASTER_LINE];
       float laser_raster_intensity_factor;
